src/requirements.txt
anyio==3.6.2
asgiref==3.4.1
astroid==2.8.4
asyncpg==0.24.0
attrs==21.2.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
databases==0.5.3
exceptiongroup==1.1.2
fastapi==0.95.2
greenlet==1.1.2
h11==0.14.0
httpcore==0.17.0
httptools==0.3.0
httpx==0.24.0
idna==3.3
iniconfig==1.1.1
isort==5.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
more-itertools==8.10.0
packaging==21.2
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pydantic==1.8.2
pylint==2.11.1
pyparsing==2.4.7
pytest==7.4.0
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing-extensions==3.10.0.2
urllib3==1.26.18
uvicorn==0.15.0
uvloop==0.16.0
wcwidth==0.2.5
websockets==10.0
wrapt==1.13.3
aio_pika==9.3.0
pika==1.3.2